Clean up codegen. We no longer need to install the cli
