--- codegen/src/cliPath.ts
@@ -1,4 +1,0 @@
-import { tmpdir } from 'os';
-import { join } from 'path';
-
-export const cliPath = join(tmpdir(), process.platform === 'win32' ? 'xata.exe' : 'xata');
