+++ codegen/src/index.ts
@@ -1,6 +1,6 @@
 export { generate } from './codegen.js';
 export type { GenerateOptions, Language } from './codegen.js';
 export type { XataConfigSchema } from './config.js';
-export { parseConfigFile } from './parseConfigFile.js';
-export { parseSchemaFile } from './parseSchemaFile.js';
+export { parseConfigFile } from './config.js';
+export { parseSchemaFile } from './schema.js';
 export type { Column, Table, XataDatabaseSchema } from './schema.js';
