+++ codegen/src/schema.ts
@@ -1,4 +1,5 @@
 import { z } from 'zod';
+import { handleParsingError } from './errors.js';
 
 // We need to do this because of problems with Zod and recursive types https://www.npmjs.com/package/zod#recursive-types
 export type Column = {
@@ -40,3 +41,12 @@ export const xataDatabaseSchema = z.object({
 });
 
 export type XataDatabaseSchema = z.infer<typeof xataDatabaseSchema>;
+
+export const parseSchemaFile = (input: string) => {
+  try {
+    return xataDatabaseSchema.parse(JSON.parse(input));
+  } catch (err) {
+    handleParsingError(err);
+    throw err; // ^ runs process.exit(1) if not successful. If not, let's throw the error because if we don't, then this function would return its type | undefined and we'd have to optionally chain and account for potential undefined returns wherever we use it.
+  }
+};
